visit DTOs

+++ src/main/java/com/wora/waiting_room/entities/Visit.java
@@ -25,7 +25,7 @@ public class Visit {
     private Status status;
 
     @Min(value = 1, message = "Priority must be at least 1")
-    private Integer priority;
+    private byte priority;
 
     @Min(value = 1, message = "Estimated processing time must be positive")
     private Integer estimatedProcessingTime;
